Render elements with function instead of impls
