--- luda-editor/luda-editor-client/src/app/sequence_list/load_button.rs
@@ -1,22 +1,0 @@
-use super::SequenceList;
-use crate::app::sequence_list::{
-    events::SequenceListEvent, rounded_rectangle::RoundedRectangleColor,
-};
-use namui::{render, RenderingTree, Wh};
-
-impl SequenceList {
-    pub fn render_load_button(&self, wh: Wh<f32>, path: &String) -> RenderingTree {
-        render![
-            self.render_rounded_rectangle(wh, RoundedRectangleColor::Blue)
-                .attach_event(move |builder| {
-                    let path = path.clone();
-                    builder.on_mouse_down(move |_| {
-                        let path = path.clone();
-                        namui::event::send(SequenceListEvent::SequenceLoadEvent { path });
-                    })
-                })
-                .with_mouse_cursor(namui::MouseCursor::Pointer),
-            self.render_button_text(wh, "Load".to_string())
-        ]
-    }
-}
