+++ luda-editor/luda-editor-client/src/app/sequence_list/reload_titles_button.rs
@@ -1,22 +1,20 @@
-use super::SequenceList;
 use crate::app::sequence_list::{
-    events::SequenceListEvent, rounded_rectangle::RoundedRectangleColor,
+    common::{render_button_text, render_rounded_rectangle, RoundedRectangleColor},
+    events::SequenceListEvent,
 };
 use namui::{render, RenderingTree, Wh};
 
-impl SequenceList {
-    pub fn render_reload_titles_button(&self, wh: Wh<f32>) -> RenderingTree {
-        render![
-            self.render_rounded_rectangle(wh, RoundedRectangleColor::Blue)
-                .attach_event(move |builder| {
-                    builder.on_mouse_down(move |_| {
-                        namui::event::send(
-                            SequenceListEvent::SequenceReloadTitlesButtonClickedEvent {},
-                        );
-                    })
+pub fn render_reload_titles_button(wh: Wh<f32>) -> RenderingTree {
+    render![
+        render_rounded_rectangle(wh, RoundedRectangleColor::Blue)
+            .attach_event(move |builder| {
+                builder.on_mouse_down(move |_| {
+                    namui::event::send(
+                        SequenceListEvent::SequenceReloadTitlesButtonClickedEvent {},
+                    );
                 })
-                .with_mouse_cursor(namui::MouseCursor::Pointer),
-            self.render_button_text(wh, "Reload titles".to_string())
-        ]
-    }
+            })
+            .with_mouse_cursor(namui::MouseCursor::Pointer),
+        render_button_text(wh, "Reload titles".to_string())
+    ]
 }
